Milhoes de ajustes

+++ Sistema de Solicitações C#/Request System/Request System/Forms/Notas Fiscais/Notas_Fiscais.cs
@@ -73,8 +73,8 @@ namespace Request_System
         private void TXT_Filter_CNPJ_TextChanged(object sender, EventArgs e)
         {
             TXT_Filter_CNPJ.BackColor = Color.White;
-            int valor;
-            if (Int32.TryParse(TXT_Filter_CNPJ.Text, out valor) || TXT_Filter_CNPJ.Text == "")
+            long valor;
+            if (Int64.TryParse(TXT_Filter_CNPJ.Text, out valor) || TXT_Filter_CNPJ.Text == "")
             {
                 BTN_Filter.PerformClick();
             }
@@ -94,7 +94,7 @@ namespace Request_System
         }
 
         private void BTN_Edit_Click_1(object sender, EventArgs e)
-        {
+        {            
             numeroNFe = Convert.ToInt32(GRID_NFes_View.CurrentRow.Cells[0].Value.ToString());
             serieNFe = Convert.ToInt32(GRID_NFes_View.CurrentRow.Cells[1].Value.ToString());
             emissionDateNFe = Convert.ToDateTime(GRID_NFes_View.CurrentRow.Cells[2].Value.ToString());
